Impl place distance computation logic

+++ clientCore/src/main/java/com/mark/zumo/client/core/entity/Store.java
@@ -28,9 +28,9 @@ public class Store {
     @SerializedName(Schema.name) @ColumnInfo(name = Schema.name)
     public final String name;
     @SerializedName(Schema.latitude) @ColumnInfo(name = Schema.latitude)
-    public final long latitude;
+    public final double latitude;
     @SerializedName(Schema.longitude) @ColumnInfo(name = Schema.longitude)
-    public final long longitude;
+    public final double longitude;
     @SerializedName(Schema.coverImageUrl) @ColumnInfo(name = Schema.coverImageUrl)
     public final String coverImageUrl;
     @SerializedName(Schema.thumbnailUrl) @ColumnInfo(name = Schema.thumbnailUrl)
@@ -38,7 +38,7 @@ public class Store {
     @SerializedName(Schema.fcmToken) @ColumnInfo(name = Schema.fcmToken)
     public final String fcmToken;
 
-    public Store(@NonNull final String uuid, final String name, final long latitude, final long longitude, final String coverImageUrl, final String thumbnailUrl, final String fcmToken) {
+    public Store(@NonNull final String uuid, final String name, final double latitude, final double longitude, final String coverImageUrl, final String thumbnailUrl, final String fcmToken) {
         this.uuid = uuid;
         this.name = name;
         this.latitude = latitude;
